Second commit.
Hey, it's starting to look more like a website now! There's something a
little weird going on with the CSS at the moment; on the 'bandlist'
page, the main content page seems to be floating to the left. It's, uh,
not supposed to do that. That said, the CSS is far from final. For
instance, the nav-bar should look fairly different by the time we're
done with it.

Oh yeah, I added some comments too.