Third commit.
I tidied up the album cover art names a bit. In addition, I added a dump
of the SQL database I'm using. If you want to import the database to
MySQL or something, be warned that it has 'DROP' statements in it, so if
you have your own MusicMashup database already, it'll overwrite any
tables and related data that you've already got.