Eighth commit
I lied. I've found a (hacky?) workaround for the issue I was having
before; it involves having a hidden input field which has a certain
number as a value. This value is checked when the file is loaded, and
the appropriate function is called. Currently the final step of the
registration process is buggy. Not sure why at the moment; it seems to
think a variable I send to it via POST doesn't exist, and my syntax for
the 'UPDATE' SQL query is wrong.