Tenth commit
Hooray! It's the tenth commit, and it's a big one! On the various 'list'
pages, you can now click on an item to view it in details. This means
that tracks, albums, artists, and users now more or less have thier own
page.

I've also implemented user registration, along with logging in and out.
One weird thing with logging in/out; when you first log in or out, the
system doesn't seem to register it until you load a new page. It's
probably easily fixable, but worth noting.

The only vital features for the main website are favouriting and the RSS
feed. Favouriting should be a doodle. We'll see about the RSS.

In other news, it should now be possible to start implementing the
widgets! I'm personally going to make a twitter widget. The others are
currently a mystery to me!