RSS, Twitter, comments
So, yeah, I broke my naming convention last update, which is irritating.
Might as well be a bit more descriptive.  I've updated Twitter so that
it caches content now. annoyingly, it only works half of the time now.
If need be, I can roll it back. I've also added the first draft of the
RSS, and It sure does produce a valid document! It doesn't quite display
all of the favourites it should (Read: It displays two), but it's on the
right track.

I've also added comments for most of the files. I think they're
sufficient; my code is pretty self-explanatory for the most part. Oh
yeah, I also added new lines, so the php output doesn't just flow in one
big line anymore.

So, what's left to do, other than making sure that the twitter widget
and RSS feed actually, y'know, work? We need to tidy up the relational
database diagram so that it matches the actual database, and fill out
the XML metadata file, so we don't automatically fail the module. We
could also implement track sorting if we have the time. Just the tracks
for the time-being, though; that's all the spec really wants, and
implementing it for every 'item' has to potential to be a nightmare, and
we only have a couple of days left.

That said, we're pretty much done, so let's give ourselves a big pat on
the back!

-Andrew