Released!
All of the widgets have been implemented, as has the ability to sort
tracks, albums and artists. Fair warning, the physical database in
database is incorrect, and will not work with this code; unfortuanately,
I cannot access the correct database at this time; I wil do this at a
later date. That said, all of the code is there, and works as expected
when you have the right database.